fixes for review
