+++ apps/etl-func/src/adapters/config.ts
@@ -9,7 +9,7 @@ export const configSchema = z.object({
   }),
   pdvTokenizer: z.object({
     apiKey: z.string().min(1),
-    baseUrl: z.string().min(1),
+    baseUrl: z.string().url(),
   }),
 });
