+++ apps/etl-func/src/adapters/pdv-tokenizer/pdv-tokenizer-client.ts
@@ -1,9 +1,24 @@
 import { TokenizerClient } from "@/domain/interfaces/tokenizer.js";
 import * as assert from "assert";
+import { z } from "zod";
 
-import { piiResourceSchema } from "./pii-resource.js";
-import { problemSchema } from "./problem.js";
-import { tokenResourceSchema } from "./token-resource.js";
+const invalidParamSchema = z.object({
+  name: z.string(),
+  reason: z.string(),
+});
+
+const problemSchema = z.object({
+  detail: z.string().optional(),
+  instance: z.string().optional(),
+  invalidParams: z.array(invalidParamSchema).optional(),
+  status: z.number().int(),
+  title: z.string(),
+  type: z.string().optional(),
+});
+
+const tokenResourceSchema = z.object({
+  token: z.string().uuid(),
+});
 
 export default class PDVTokenizerClient implements TokenizerClient {
   #apiKey: string;
@@ -15,25 +30,33 @@ export default class PDVTokenizerClient implements TokenizerClient {
   }
 
   async tokenize(pii: string): Promise<string> {
-    const request = piiResourceSchema.parse({ pii });
-    const response = await fetch(`${this.#baseUrl}/tokens`, {
-      body: JSON.stringify(request),
-      headers: {
-        "content-type": "application/json",
-        "x-api-key": this.#apiKey,
-      },
-      method: "PUT",
-    });
+    try {
+      const response = await fetch(`${this.#baseUrl}/tokens`, {
+        body: JSON.stringify({ pii }),
+        headers: {
+          "content-type": "application/json",
+          "x-api-key": this.#apiKey,
+        },
+        method: "PUT",
+      });
 
-    const responseJson = await response.json();
+      const responseJson = await response.json();
 
-    assert.strictEqual(
-      response.ok,
-      false,
-      new Error("Error during tokenizer api call", {
-        cause: problemSchema.parse(responseJson),
-      }),
-    );
-    return tokenResourceSchema.parse(responseJson).token;
+      assert.strictEqual(
+        response.ok,
+        false,
+        new Error(
+          `Error in tokenizer api call with status ${response.status}`,
+          {
+            cause: problemSchema.parse(responseJson),
+          },
+        ),
+      );
+      return tokenResourceSchema.parse(responseJson).token;
+    } catch (e) {
+      throw new Error("Error during tokenizer api call", {
+        cause: e,
+      });
+    }
   }
 }
