--- apps/etl-func/src/adapters/pdv-tokenizer/pii-resource.ts
@@ -1,4 +1,0 @@
-import { z } from "zod";
-
-export const piiResourceSchema = z.object({ pii: z.string().min(1) });
-export type PiiResource = z.TypeOf<typeof piiResourceSchema>;
