--- apps/etl-func/src/adapters/pdv-tokenizer/token-resource.ts
@@ -1,6 +1,0 @@
-import { z } from "zod";
-
-export const tokenResourceSchema = z.object({
-  token: z.string().uuid(),
-});
-export type TokenResource = z.TypeOf<typeof tokenResourceSchema>;
